Version 4 (#72)
* Restructuring solution for v4

* Restructure MySQL

* Fixing MSSQL + bump to .NET6

* MSSQL Select tests

* MSSQL delete tests

* MSSQL insert tests

* MSSQL Issues tests

* MSSQL update tests

* MSSQL Aggregates Tests

* MySQL tests

* Adding PostgreSQL tests

* Support for OPTION(RECOMPILE) #70

* Support for ilike in postgresql #65

* Releasing v4

+++ tests/Dapper.FSharp.Tests/MSSQL/Database.fs
@@ -1,19 +1,33 @@
 ﻿module Dapper.FSharp.Tests.MSSQL.Database
 
+open Dapper.FSharp
 open Dapper.FSharp.Tests.Database
 open Dapper.FSharp.Tests.Extensions
 open System.Data
-open FSharp.Control.Tasks
+open Microsoft.Data.SqlClient
+open Microsoft.Extensions.Configuration
 
-let init (conn:IDbConnection) =
+let getConnection () =
+    let conf = ConfigurationBuilder().AddJsonFile("settings.json").Build()
+    new SqlConnection(conf.["mssqlConnectionString"])
+
+let mutable isAlreadyInitialized = false
+
+let safeInit (conn:IDbConnection) =
     task {
-        do! DbName |> sprintf "DROP DATABASE IF EXISTS %s;" |> conn.ExecuteIgnore
-        do! DbName |> sprintf "CREATE DATABASE %s;" |> conn.ExecuteIgnore
-        conn.Open()
-        conn.ChangeDatabase DbName
-        do! TestSchema |> sprintf "DROP SCHEMA IF EXISTS %s;" |> conn.ExecuteIgnore
-        do! TestSchema |> sprintf "CREATE SCHEMA %s;" |> conn.ExecuteIgnore
-    } |> Async.AwaitTask |> Async.RunSynchronously
+        if isAlreadyInitialized |> not then
+            
+            do! DbName |> sprintf "DROP DATABASE IF EXISTS %s;" |> conn.ExecuteIgnore
+            do! DbName |> sprintf "CREATE DATABASE %s;" |> conn.ExecuteIgnore
+            conn.Open()
+            conn.ChangeDatabase DbName
+            do! TestSchema |> sprintf "DROP SCHEMA IF EXISTS %s;" |> conn.ExecuteIgnore
+            do! TestSchema |> sprintf "CREATE SCHEMA %s;" |> conn.ExecuteIgnore
+            isAlreadyInitialized <- true
+            OptionTypes.register()
+    }
+    |> Async.AwaitTask
+    |> Async.RunSynchronously
 
 module Persons =
 
@@ -170,22 +184,6 @@ module Issues =
                 return ()
             }
 
-open Dapper.FSharp.MSSQL
-
-let getCrud (conn:IDbConnection) =
-    { new ICrudOutput with
-        member x.SelectAsync<'a> (q, cancellationToken) = conn.SelectAsync<'a>(q, ?cancellationToken = cancellationToken)
-        member x.SelectAsync<'a,'b> q = conn.SelectAsync<'a,'b>(q)
-        member x.SelectAsync<'a,'b,'c> q = conn.SelectAsync<'a,'b,'c>(q)
-        member x.SelectAsyncOption<'a,'b> q = conn.SelectAsyncOption<'a,'b>(q)
-        member x.SelectAsyncOption<'a,'b,'c> q = conn.SelectAsyncOption<'a,'b,'c>(q)
-        member x.InsertAsync<'a> (q, cancellationToken) = conn.InsertAsync<'a>(q, ?cancellationToken = cancellationToken)
-        member x.DeleteAsync (q, cancellationToken) = conn.DeleteAsync(q, ?cancellationToken = cancellationToken)
-        member x.UpdateAsync (q, cancellationToken) = conn.UpdateAsync(q, ?cancellationToken = cancellationToken)
-        member x.InsertOutputAsync q = conn.InsertOutputAsync(q)
-        member x.DeleteOutputAsync q = conn.DeleteOutputAsync(q)
-        member x.UpdateOutputAsync q = conn.UpdateOutputAsync(q)
-    }
 
 let getInitializer (conn:IDbConnection) =
     { new ICrudInitializer with
